% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{spdf_close}
\alias{spdf_close}
\title{Close a SPDF document}
\usage{
spdf_close(doc)
}
\arguments{
\item{doc}{A \code{spdf_document}.}
}
\value{
\code{doc}, invisibly.
}
\description{
Close a SPDF document
}
